requirements.txt
numpy==1.20.3
pandas==1.3.4
scikit-learn==1.0.2
seaborn==0.11.2
snowflake-connector-python==2.7.1
matplotlib==3.4.3
plotly==5.5.0
streamlit==1.5.0